requirements.txt
google-adk
litellm
google_cloud_aiplatform==1.98.0
protobuf==6.31.1
python-dotenv==1.1.1